+++ src/render/samplers/JitteredSampler.cpp
@@ -6,6 +6,15 @@
 using namespace render;
 using namespace std;
 
+Vector2d JitteredSampler::sampleForDimension(int sampleIndex, uint64_t pixelHash,
+                                             uint64_t dimension) const {
+  if (!isPathTracingDimension(dimension)) {
+    return Sampler::sampleForDimension(sampleIndex, pixelHash, dimension);
+  }
+
+  return offsetScrambledPathDimensionSample(sampleIndex, pixelHash, dimension);
+}
+
 std::shared_ptr<SampleStream> JitteredSampler::sharedStream(int sampleIndex,
                                                             uint64_t pixelHash) const {
   return sharedSamplerBackedStream(sampleIndex, pixelHash);